restructure modules, add tests for modelfunction builder, fix bugs
